// src/typescript/simplestamp/models/simplestamp/v1/attestation.ts
export interface Attestation {
  /** The fully qualified URL of the calendar server the stamping was made on. */
  calendarUrl: string;
  /** UNIX timestamp of when the call to the calendar server was made. */
  submitted: number;
  /**
   * The sequence of operations the calendar server responded with,
   * performed on the original hash to get the timestamp URL on the
   * server for status.
   */
  operations: Operation[];
  status: AttestationStatus;
  /** If the attestation indicates it is on chain */
  blockHeight: number;
  /** The merkle root of the block containing the transaction with verification */
  blockMerkleRoot: Uint8Array;
  /**
   * The merkle root of all of the hashes committed to the Open Timestamps server.
   * This is committed to the transaction on chain as an OP_RETURN
   */
  timestampMerkleRoot: Uint8Array;
  /**
   * The transaction ID of the on chain transaction holding the merkle root of
   * the calendar server data being attested.
   */
  transactionId: Uint8Array;
}

function createBaseAttestation(): Attestation {
  return {
    calendarUrl: "",
    submitted: 0,
    operations: [],
    status: 0,
    blockHeight: 0,
    blockMerkleRoot: new Uint8Array(0),
    timestampMerkleRoot: new Uint8Array(0),
    transactionId: new Uint8Array(0),
  };
}

export const Attestation: MessageFns<Attestation> = {
  encode(message: Attestation, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.calendarUrl !== "") {
      writer.uint32(10).string(message.calendarUrl);
    }
    if (message.submitted !== 0) {
      writer.uint32(16).uint32(message.submitted);
    }
    for (const v of message.operations) {
      Operation.encode(v!, writer.uint32(26).fork()).join();
    }
    if (message.status !== 0) {
      writer.uint32(32).int32(message.status);
    }
    if (message.blockHeight !== 0) {
      writer.uint32(40).uint32(message.blockHeight);
    }
    if (message.blockMerkleRoot.length !== 0) {
      writer.uint32(50).bytes(message.blockMerkleRoot);
    }
    if (message.timestampMerkleRoot.length !== 0) {
      writer.uint32(58).bytes(message.timestampMerkleRoot);
    }
    if (message.transactionId.length !== 0) {
      writer.uint32(66).bytes(message.transactionId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Attestation {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAttestation();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.calendarUrl = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.submitted = reader.uint32();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.operations.push(Operation.decode(reader, reader.uint32()));
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.status = reader.int32() as any;
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.blockHeight = reader.uint32();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.blockMerkleRoot = reader.bytes();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.timestampMerkleRoot = reader.bytes();
          continue;
        }
        case 8: {
          if (tag !== 66) {
            break;
          }

          message.transactionId = reader.bytes();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<Attestation>, I>>(base?: I): Attestation {
    return Attestation.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Attestation>, I>>(object: I): Attestation {
    const message = createBaseAttestation();
    message.calendarUrl = object.calendarUrl ?? "";
    message.submitted = object.submitted ?? 0;
    message.operations = object.operations?.map((e) => Operation.fromPartial(e)) || [];
    message.status = object.status ?? 0;
    message.blockHeight = object.blockHeight ?? 0;
    message.blockMerkleRoot = object.blockMerkleRoot ?? new Uint8Array(0);
    message.timestampMerkleRoot = object.timestampMerkleRoot ?? new Uint8Array(0);
    message.transactionId = object.transactionId ?? new Uint8Array(0);
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}