// src/typescript/simplestamp/models/simplestamp/v1/status.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.11.8
//   protoc               v7.35.0
// source: simplestamp/v1/status.proto

/* eslint-disable */

export const protobufPackage = "simplestamp.v1";

/**
 * SimpleStamp data models for storing attestations.
 *
 * License: Unlicense https://unlicense.org
 */

/** The status of the attestation as reported by the calendar server. */
export enum AttestationStatus {
  ATTESTATION_STATUS_INVALID = 0,
  ATTESTATION_STATUS_UNKNOWN = 1,
  ATTESTATION_STATUS_PENDING = 2,
  ATTESTATION_STATUS_BITCOIN = 3,
  ATTESTATION_STATUS_LITECOIN = 4,
  UNRECOGNIZED = -1,
}