// src/typescript/simplestamp/models/simplestamp/v1/identity.ts
export interface Identity {
  /** 2 letter country code/name */
  countryCode: string;
  /** Full name of state or province */
  state: string;
  /** City or locality name */
  city: string;
  /** Organization name */
  organization: string;
  /** Organization unit or section */
  section: string;
  /** Common name, fully qualified host name */
  commonName: string;
  /** Email address */
  email: string;
  /** Full name of the person */
  fullName: string;
}

function createBaseIdentity(): Identity {
  return {
    countryCode: "",
    state: "",
    city: "",
    organization: "",
    section: "",
    commonName: "",
    email: "",
    fullName: "",
  };
}

export const Identity: MessageFns<Identity> = {
  encode(message: Identity, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.countryCode !== "") {
      writer.uint32(10).string(message.countryCode);
    }
    if (message.state !== "") {
      writer.uint32(18).string(message.state);
    }
    if (message.city !== "") {
      writer.uint32(26).string(message.city);
    }
    if (message.organization !== "") {
      writer.uint32(34).string(message.organization);
    }
    if (message.section !== "") {
      writer.uint32(42).string(message.section);
    }
    if (message.commonName !== "") {
      writer.uint32(50).string(message.commonName);
    }
    if (message.email !== "") {
      writer.uint32(58).string(message.email);
    }
    if (message.fullName !== "") {
      writer.uint32(66).string(message.fullName);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Identity {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseIdentity();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.countryCode = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.state = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.city = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.organization = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.section = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.commonName = reader.string();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.email = reader.string();
          continue;
        }
        case 8: {
          if (tag !== 66) {
            break;
          }

          message.fullName = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<Identity>, I>>(base?: I): Identity {
    return Identity.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Identity>, I>>(object: I): Identity {
    const message = createBaseIdentity();
    message.countryCode = object.countryCode ?? "";
    message.state = object.state ?? "";
    message.city = object.city ?? "";
    message.organization = object.organization ?? "";
    message.section = object.section ?? "";
    message.commonName = object.commonName ?? "";
    message.email = object.email ?? "";
    message.fullName = object.fullName ?? "";
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}