// src/typescript/simplestamp/models/simplestamp/v1/location.ts
export interface Location {
  /** Latitude and longitude in the decimal notation format */
  latitude: number;
  longitude: number;
  /** Elevation above or below sea level in meters */
  altitude: number;
  /** If the device determining location reports the accuracy */
  accuracyMeters: number;
  /**
   * Relative angle at which the trajectory is headed, from 0 to 360, with
   * 0 being true north and degrees being clockwise
   */
  direction: number;
  /** Speed in meters/second */
  velocity: number;
}

function createBaseLocation(): Location {
  return { latitude: 0, longitude: 0, altitude: 0, accuracyMeters: 0, direction: 0, velocity: 0 };
}

export const Location: MessageFns<Location> = {
  encode(message: Location, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.latitude !== 0) {
      writer.uint32(9).double(message.latitude);
    }
    if (message.longitude !== 0) {
      writer.uint32(17).double(message.longitude);
    }
    if (message.altitude !== 0) {
      writer.uint32(25).double(message.altitude);
    }
    if (message.accuracyMeters !== 0) {
      writer.uint32(37).float(message.accuracyMeters);
    }
    if (message.direction !== 0) {
      writer.uint32(45).float(message.direction);
    }
    if (message.velocity !== 0) {
      writer.uint32(53).float(message.velocity);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Location {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLocation();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 9) {
            break;
          }

          message.latitude = reader.double();
          continue;
        }
        case 2: {
          if (tag !== 17) {
            break;
          }

          message.longitude = reader.double();
          continue;
        }
        case 3: {
          if (tag !== 25) {
            break;
          }

          message.altitude = reader.double();
          continue;
        }
        case 4: {
          if (tag !== 37) {
            break;
          }

          message.accuracyMeters = reader.float();
          continue;
        }
        case 5: {
          if (tag !== 45) {
            break;
          }

          message.direction = reader.float();
          continue;
        }
        case 6: {
          if (tag !== 53) {
            break;
          }

          message.velocity = reader.float();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<Location>, I>>(base?: I): Location {
    return Location.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Location>, I>>(object: I): Location {
    const message = createBaseLocation();
    message.latitude = object.latitude ?? 0;
    message.longitude = object.longitude ?? 0;
    message.altitude = object.altitude ?? 0;
    message.accuracyMeters = object.accuracyMeters ?? 0;
    message.direction = object.direction ?? 0;
    message.velocity = object.velocity ?? 0;
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}