// src/typescript/simplestamp/models/simplestamp/v1/timestamp.ts
export interface SimpleStamp {
  /**
   * The binary hash of the data this timestamp is for.
   * Most likely the SHA256 of a file.
   */
  hash: Uint8Array;
  /**
   * The random data that is appended to the hash on creation.  This results in
   * stamping the same file multiple times and getting a unique stamp. The value
   * submitted to the server is then: SHA256(CONCAT(hash, nonce))
   */
  nonce: Uint8Array;
  /** The attestations made */
  attestations: Attestation[];
  /** The UNIX timestamp of when this was originally created */
  created: number;
  /**
   * URL, file name, or general description for where the data
   * used to generate the hash came from.
   */
  source: string;
  /** Identity of the organization and/or person stamping */
  identity:
    | Identity
    | undefined;
  /** Where the timestamp was created and it's trajectory if moving */
  location:
    | Location
    | undefined;
  /** Assorted free form details for the creator to describe the purpose or intent */
  description: string;
  /**
   * Cryptographic signature of the hash sent to digest. This lets the person making
   * the Timestamp use a private key to sign, so others can validate with their
   * public key.
   */
  signature: Uint8Array;
  /**
   * Unique identifier for the key used to sign. Open ended, to help identity which key was
   * used. Should not contain anything from the key itself.
   */
  keyId: string;
}

function createBaseSimpleStamp(): SimpleStamp {
  return {
    hash: new Uint8Array(0),
    nonce: new Uint8Array(0),
    attestations: [],
    created: 0,
    source: "",
    identity: undefined,
    location: undefined,
    description: "",
    signature: new Uint8Array(0),
    keyId: "",
  };
}

export const SimpleStamp: MessageFns<SimpleStamp> = {
  encode(message: SimpleStamp, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.hash.length !== 0) {
      writer.uint32(10).bytes(message.hash);
    }
    if (message.nonce.length !== 0) {
      writer.uint32(18).bytes(message.nonce);
    }
    for (const v of message.attestations) {
      Attestation.encode(v!, writer.uint32(26).fork()).join();
    }
    if (message.created !== 0) {
      writer.uint32(32).uint32(message.created);
    }
    if (message.source !== "") {
      writer.uint32(42).string(message.source);
    }
    if (message.identity !== undefined) {
      Identity.encode(message.identity, writer.uint32(50).fork()).join();
    }
    if (message.location !== undefined) {
      Location.encode(message.location, writer.uint32(58).fork()).join();
    }
    if (message.description !== "") {
      writer.uint32(66).string(message.description);
    }
    if (message.signature.length !== 0) {
      writer.uint32(74).bytes(message.signature);
    }
    if (message.keyId !== "") {
      writer.uint32(82).string(message.keyId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SimpleStamp {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSimpleStamp();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.hash = reader.bytes();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.nonce = reader.bytes();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.attestations.push(Attestation.decode(reader, reader.uint32()));
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.created = reader.uint32();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.source = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.identity = Identity.decode(reader, reader.uint32());
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.location = Location.decode(reader, reader.uint32());
          continue;
        }
        case 8: {
          if (tag !== 66) {
            break;
          }

          message.description = reader.string();
          continue;
        }
        case 9: {
          if (tag !== 74) {
            break;
          }

          message.signature = reader.bytes();
          continue;
        }
        case 10: {
          if (tag !== 82) {
            break;
          }

          message.keyId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<SimpleStamp>, I>>(base?: I): SimpleStamp {
    return SimpleStamp.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SimpleStamp>, I>>(object: I): SimpleStamp {
    const message = createBaseSimpleStamp();
    message.hash = object.hash ?? new Uint8Array(0);
    message.nonce = object.nonce ?? new Uint8Array(0);
    message.attestations = object.attestations?.map((e) => Attestation.fromPartial(e)) || [];
    message.created = object.created ?? 0;
    message.source = object.source ?? "";
    message.identity = (object.identity !== undefined && object.identity !== null)
      ? Identity.fromPartial(object.identity)
      : undefined;
    message.location = (object.location !== undefined && object.location !== null)
      ? Location.fromPartial(object.location)
      : undefined;
    message.description = object.description ?? "";
    message.signature = object.signature ?? new Uint8Array(0);
    message.keyId = object.keyId ?? "";
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}