// src/typescript/simplestamp/models/simplestamp/v1/operation.ts
export enum OperationType {
  OPERATION_TYPE_UNKNOWN = 0,
  OPERATION_TYPE_ATTESTATION = 1,
  OPERATION_TYPE_SHA1 = 2,
  OPERATION_TYPE_RIPEMD160 = 3,
  OPERATION_TYPE_SHA256 = 8,
  OPERATION_TYPE_APPEND = 240,
  OPERATION_TYPE_PREPEND = 241,
  OPERATION_TYPE_REVERSE = 242,
  OPERATION_TYPE_HEXLIFY = 243,
  UNRECOGNIZED = -1,
}